+++ automerge-protocol/src/serde_impls/change_hash.rs
