+++ automerge-protocol/src/serde_impls/mod.rs
